refactor(search-api): adopt src layout and fix packaging
