+++ services/search-api/app/__init__.py
@@ -1,3 +1,2 @@
-from .main import app
-
-__all__ = ["app"]
+"""Backward compatibility package mapping to search_api.app."""
+from search_api.app import *  # noqa: F401,F403
