+++ services/search-api/src/search_api/__init__.py
@@ -1,0 +1,1 @@
+"""Search API package."""
