+++ services/search-api/src/search_api/_shared/__init__.py
@@ -1,0 +1,1 @@
+# Shared utilities for InfoTerminal services
