+++ services/search-api/tests/conftest.py
@@ -1,19 +1,29 @@
-import os, pathlib, pytest, sys, importlib
+import os
+import pathlib
+import sys
+import importlib
+import pytest
 from httpx import AsyncClient, ASGITransport
 
 os.environ.setdefault("OTEL_SDK_DISABLED", "true")
 os.environ.setdefault("IT_JSON_LOGS", "1")
 os.environ.setdefault("IT_ENV", "test")
 os.environ.setdefault("IT_LOG_SAMPLING", "")
+
 SERVICE_DIR = pathlib.Path(__file__).resolve().parents[1]
+SRC_DIR = SERVICE_DIR / "src"
+sys.path.insert(0, str(SRC_DIR))
 sys.path.insert(0, str(SERVICE_DIR))
-import app.main as app_main  # type: ignore
+
+import search_api.app.main as app_main  # type: ignore
 sys.modules.setdefault("app.main", app_main)
 app = app_main.app
+
 
 @pytest.fixture(scope="session")
 def anyio_backend():
     return "asyncio"
+
 
 @pytest.fixture(scope="session")
 def test_settings():
@@ -24,6 +34,7 @@ def test_settings():
     yield
     for k in ("ALLOW_TEST_MODE", "IT_JSON_LOGS", "IT_ENV", "IT_LOG_SAMPLING", "IT_OTEL"):
         os.environ.pop(k, None)
+
 
 @pytest.fixture
 async def client(test_settings):
